Add global option to disable/enable the API
Added a new section to the preferences for global settings, which affect
all tournaments on the instance of Tabbycat. This addition is based on
issue #685. As a first setting, a toggle for the API was added.

The API activation is done through permissions; if disabled, a 401 is
thrown with a "disabled" message.