Talk about API in documentation
This commit adds a page in the documentation about the API, what it is
and how it's used, as well as information about authentication and
authorization. The latter focuses on the access limitations and
customizing access.

A passage about the API was also added to the import page, as the API
provides endpoints to create participants, so is another way for import.