Remove more slashes from API endpoint URLs
This commit does a couple of things.

Firstly, it creates a root API view which links to the v1 root; removing
slashes from the roots. Services shouldn't need to know what version
we're currently on. Then, it imbriques the API views inside `v1`,
which is moved to the API URLs file.

Then, it moves the API URLs inclusion above that of tournaments to
prevent redirecting to a tournament with slug "api".

Finally, it silences the missing trailing slash warning globally. As the
API URLs don't use it as explained, the warnings should be ignored. The
silencing is global just because there is no granularity. (Django26472)