Add conflicts to serializers
This commit adds adjudicator, institution, and team conflicts to the
Adjudicator viewset as well as institutional conflicts for the Team
viewset, which are only visible as administrator. For this, ManyToMany
fields were added/adapted to the models as a shortcut.

The self-ManyToMany should not be a concern as "duplicate" conflicts are
discounted in allocations. The serializer only takes targets from the
source, and not also the reverse.

Also fixed an oversight which disabled the speaker serializer when not
nested.