Re-order error reporting venues
When a 500 (Internal Server) Error occurs, a page is displayed prompting
users to report the problem. It includes an order of precedence for the
best place to report. However, the specialities of the services are not
and cannot really be exposed for self-triage. Whereas GitHub tasks are
more for code issues, the Facebook group is more general. As non-code
(like hosting problems) are more common, the Facebook group should be
preferred.