Make set_score silently fail if adj not in debate
This makes `VotingDebateResult.set_score(adjudicator, *args)` fail
silently if there isn't a scoresheet for `adjudicator`, which
(should) happen if the adjudicator isn't assigned to that debate.
It logs an exception, so the failure isn't completely silent, but
it doesn't alert the user immediately.

Relates to #922. I'm not yet sure if making this silently fail
might cause problems later.