Feedback API: Move participant testing to validation
This commit moves the verification that the source of the feedback was
in the debate specified, and adds an error message if it's not the case.
It also creates a new check that the target was also in the debate.

With these checks, we can "move" the values to proper keys so that
they don't have to be treated in save().