Synchronously check in/out in auxiliary methods
It can happen that the checkins consumer is disconnected, and if that's
the case, methods of performing check ins/outs that don't involve
directly opening the websocket may silently fail. This happens with
private URLs and API features. So, instead of having the websocket
perform the action, these pages will perform it themselves and broadcast
the result, which would be picked up if the connection is active.