Use ArrayField for "Select Multiple" and choices
For Adjudicator feedback questions and answers. This creates a new table
for Select Multiple answers with an ArrayField for the answers given.

The question model is also adapted to change the choices field to an
array as well. Both changes remove the need for delimiters between
choices. A migration is also added which converts the Q&A types.