Process queryset metrics all at once
This commit separates the handling of queryset annotations and other
metric annotators in standings, so that the queries do not need to be
re-evaluated. For that, metric keys are used as column names. The
distinct() method was also removed as it wasn't necessary.

All queryset aggregation annotators are split in two parts; annotating
the queryset, then annotating standings.

Trimmed mean has been moved down to after the number of speeches metric.
As column names can't be specified, the count is subclassed with a
different key. Draw strength now uses getattr() for the opponent metric
key.