Complicate handling of emoji in API
Blank emoji field should be allowed if no emoji is to be given to the
team. However, DB constraints mean they have to be null in that case.

Null values are currently used to indicate they should be automatically
generated, so now the check is first if null, to generate. After, if
blank, change to null. The validation is thus removed.