Re-work conditionals for names in registration card
There's no harm in showing the full name if the short name is shown,
even when identical. That conditional is thus removed. Thus, it should
show on private URL pages.

If one of the names cannot be shown, the emoji should be moved to the
full name to still be shown.