Protect records of user actions
This commit sets the user foreign key fields on models to PROTECT, which
means that associated users cannot be deleted until the objects are
either deleted as well, or manually null'd. This is important as a last
check that the user is to be deleted, and the consequences of doing so
(they shouldn't usually be deleted).

This commit also creates the migration renaming "email" for
participants.