Restrict debate status options for postponements
If postponements are disabled (which they are by default, and are not
yet implemented), the drop-down for the debate status should not show
the option. Currently, choosing that option causes an error. This commit
checks for the new preference and adapts the options to that.

The new preference re-uses the old text for the preference before
deletion.