Disable contextual alternates for plain URLs
Inter (font) provides alternates for glyphs when used in various
circumstances. However, one of these alternates 'x' (U+0078) with '×'
(U+00D7) when between two digits (supposing a multiplication). This is
not the case though with URL keys when shown in the URLs in the list.

This commit disables the alternates in this case, and provides a CSS
class to use elsewhere if needed (.mixed-text).