i18n: Combine BP result messages
This commit simplifies the messages for debate results in BP (like
"%(team)s took %(place)s") so that all places use the same terminology
of "placed" rather than "took", and combining the messages to abstract
the position, using the ordinal function of Django for the ordinal.

Also, instead of testing the value of the value with a chain of ifs,
it'll just be tested for being null. As a big integer (more than 3)
would be problematic, an exception is raised, rather than hidden as
unknown.