Correspond max team names lengths to generation
This commit modifies the Team model to allow for longer team names
(short_name and long_name) as their max lengths did not correspond with
the maximum possible lengths of the strings generated for those fields.

As such, the max lengths for these fields is shown as the max length of
the institution code/name plus a space and the team [short_]reference.