Use window functions with standings rankings
This commit adds methods to Standings and rankings to take advantage of
window functions (RANK()) to rank participants if all standings metrics
are also SQL aggregations.

The rank uses RANK() with precedence keys to order by, and a count of
the number of participants with the same value for the precedence being
a Count() window partitioned by the keys. Subranks and other rankings
are partitioned by the first metric/values then ordered by the next.