API: Generate URL key on request
This commit sets the URL key on Speaker and Adjudicator endpoints on
creation of a participant like as done with team emojis. If the field
is not present or null, the key will be generated, but if it's an empty
string, it will be set as null and not automatically generated.

This is done with the privateurls utils, so the generation is performed
after the creation of the object, but the checks are done beforehand.