Create Results API endpoints
This commit creates API endpoints that can be used to get and create
debate results. Result endpoints are under each debate (pairing), and
can be filtered to just the confirmed ballots. When public, only the
confirmed ballot is shown.

The base for the results is in a new class ResultInfo, which takes a
DebateResult object and transforms the methods to retrieve info into
object attributes, that can then be read by DRF serializer fields.
Serializer subclasses were created to mirror the new DebateInfo
structure, but the save() methods bypass ResultInfo to set and save
DebateResult directly.