Add field for participant in Submission
This commit adds a new field to Submission models as a Person foreign
key. This field is then populated when using private URLs as an
alternative to IP addresses (although still collected).

With this, ballots can be more easily distinguished from which
adjudicator had submitted the ballot. This is added to the "other
ballots" list. It can also be used to find which team member submitted
feedback, although not shown in any template.