Enable public pairing endpoints with result release
This commit adds other preference checks to the pairings API viewset, so
that the viewset is enabled for the public when the results for that
round are released, as the results table also shows the pairings.

This also expands the data available in the pairings views with the
status and importance(s) fields. These are always hidden from the
public.