Avoid re-fetching objects for notifications
While the consumers need to be passed PKs and not the objects, the
generators which are then called can be passed objects directly. This
commit makes the generators take Tournament, Round, or Debate objects
depending on which are required. It also removes other arguments that
can be inferred from another. Select-related is used to make it more
efficient.