API: Represent Region as a string
Instead of having regions in the institutions API appear as unlinked
PKs, this change makes the regions appear as their names, which can be
retrieved and created by setting in the institutions endpoints.

The endpoints can also now be filtered by region name.

This change also prefixes the custom serializer fields instead of
listing them all in the import.