Update fixtures and add reverse migrations
This is primarily for the adjfeedback migration that splits the choice
fields into proper arrays. The fixture updating script expects backward
migrations, so one was added, with the deletion of the old objects.

With the backwards migrations, the fixtures were updated.

Also, authtokens are prevented from being included in the fixtures as
the combination of uniqueness and creation on signal makes an error as
the tests try to re-create the objects.