Add round weightings for points
This commit implements a mechanism for tapered scoring as introduced in
International Debate in March. Here, weights can be added to rounds that
act as coefficients to the sum of points team metric. Weights are
customizable as model fields, with a migration created to add it.

Weights only apply to the points metric, and not to metrics counting
the number of a rank received (number of firsts, wins, etc.)