Fix debate flags display error and form field
This fixes a couple of issues that probably arose as a result of #1525.

1. Currently, the draw page displays the characters of `str(debate.flags)`
on one line each (so "[", "]" for no flags). This commit fixes it so
that it just displays the human-readable versions of the flags, restoring
the behavior before #1525 was merged.

2. Currently, the form field for flags in the Edit Database area rejects
the value whenever one or more fields are selected. This commit modifies
the flags fields to use a new `ChoiceArrayField` that is like an
`ArrayField`, but handles choices using a `TypedMultipleChoiceField`.

The ChoiceArrayField code was taken from:
https://gist.github.com/danni/f55c4ce19598b2b345ef

See also related Django issue tracker:
https://code.djangoproject.com/ticket/27704

I couldn't figure out how to (easily) get the form field in the Edit
Database area to be taller than 2.28571rem, the height of most fields
in django-jet, without an exception made for select[multiple].