Fix "complete round" button reversal from 8a84bc613
The change in 8a84bc613ecb61383ac17412c3cb6b9d95ddd204 reversed
disabling of the "complete round" button, by removing the "== 0"
part without swapping the body of the if-else template tag. This
swaps the body to restore original behavior, while preserving
the addition of the .exists() call for efficiency.