Add "number of thirds" metric annotator
Even though a "number of thirds" annotator would be useless in normal
BP tournaments as the value can be deduced from the number of firsts/
seconds and total points, this is not the case with round weights.

The position count annotators do not take weights into consideration, so
the value of a 3rd could be the value of another ranking in different
rounds. This value can now be used as a tie-breaker.