Change tests to show null standings as last
The tests created to test standings with an extra team with no results
had the extra team at the same rank (and thus tied) with the lowest-
ranked "real" team. However, participants with no data shouldn't be
grouped with the lowest because these situations between null and 0 are
not equivalent.

The "wins" test remains as the Count() aggregate doesn't distinguish
nulls from non-true values.