Move subject/body from SentMessage to BulkNotification
This commit shifts the storage of email contents from a complete message
to just the templates used to create the message body (and subject).
With this change, the values can be shifted to the BulkNotification
table.

This shift will drastically reduce the size of the SentMessage table,
while keeping the possibility to reconstruct old messages using the
context field from SentMessage with the templates. However, most headers
will be lost with the migration. There was also a problem with messages
being sent but without SentMessage creation. I think it's due to
exceeding the maximum size of a PostgreSQL query, and this change should
help.