Use order_by() for queryset-based standings
This commit adds an order_by clause to standings based on querysets that
orders based on the ranking metrics, and then by the specified tiebreak
field (as duplicated from the StandingsInfo-based standings). This makes
the database responsible for even more, if able.

Also some code reductions were done.

Depends on Django bug 26390 to be fixed.