Add API endpoints for team breaks
This commit creates an endpoint to get the breaking teams (BreakingTeam)
for a break category, and handle the object. These endpoints are List-
based as being a list of BreakingTeam, but with adapted methods.

GET gives the team break, POST creates the break, DELETE deletes it, and
PATCH updates the remark of a team and then regenerates that break. POST
and PATCH return the break as modified.