Split motion statistics queries
The Motion/RoundMotion querysets were massive and inefficient, due to
sprawling joins, which then necessitates the use of the DISTINCT
keyword, causing massive efficiency problems. This commit breaks up the
queries in a way to avoid conflicting joins, avoiding the DISTINCT
keyword, and then recombining the annotations as attributes on the
objects.

The commit also changes the order of the Round badges to avoid reversing
the Round prefetch, which causes a reevaluation of the query.