Add error to options index for preallocated sides
This commit adds a message to the page alerts on the options index to
warn against the interaction between Intermediate 1 and Intermediate 2
without using preallocated sides. This is the subject of BACKEND-20B.

It also moves the error sent by the messages framework to the same page
for "per-adjudicator" sheets in BP, moving it to HTML with a link to the
referenced section.