Use bulk creation with own institution conflicts
This commit generalizes the addowninstitutionconflicts command as to
treat teams and adjudicators through the same function, taking the fact
that the field names are identical.

The queries are also filterable in the same way to remove the verbose
loop, replacing with a bulk creation with list comprehension. (The
conflict model can be deduced from the queryset).