Coalesce number of adjs metric to 0
If a TeamScore object has a votes_possible field of 0, the division-by-
zero error is caught and returns NULL in the query. However, when
casting the value between int and float, this causes an exception.

PostgreSQL's Coalesce function is used to use 0 if null.

This commit also cleans up NullIf as it is now built-in.

Fixes #1682
Closes BACKEND-45K