Process standings metrics without DISTINCT
The AISS team metric and well as the team points speaker metric joins
other models, and so would conflict with other metrics if there is no
DISTINCT keyword, making the values meaningless.

To counteract, a new attribute for metrics is introduced, being non-
combinable queryset metrics, where non-combinable metrics are evaluated
on-place of the queryset and first, and uses the default ranking
generators if in the precedence.