Remove note about commas from c61a98c9
This renames the migration to avoid a chain of pointless migrations.
It should work okay with databases that already have the old migration,
but it might be helpful to unmigrate back to 0010 (using commit
c61a98c9) or to manually remove the migration record for all other
adjfeedback.0011_* migrations (if any) before applying this.