Don't calculate liveness for elimination rounds
This commit removes the break category highlights from allocation views
in elimination rounds, as all the teams are live and in a specific
category.

The liveness calculations are updated so that missing team points do not
cause the liveness calculations to crash, and letting the DB sort.

Fixes BACKEND-4DQ
Fixes BACKEND-4DR