Add more stringent test for number advancing
To avoid calling round.is_last as often, the conditions that lead to
that point were expanded and analyzed (added parentheses too) to make
sure the right result class is used.

Then, for ConsensusDebateResult, the condition for having just 1 team
advance in BP now directly tests if the scoresheet is correct. So, the
.is_last method is only called for BP outrounds.

Ref #1704