Associate more motions to ballots
Currently, motions are only associated with ballot submissions directly
if the 'enable_motions' preference is enabled. This creates two paths to
get motions from debates, and ignores edge-cases for special motions but
without selection. This commit adds the motion foreign key to all
ballots regardless of the preference, added automatically if there is
only one option.

A migration is there to populate the field, but is conservative to not
prefer a motion over another and to only update where motions are null.