Associate more Venues[Categories] to tournaments
This commit makes sure to assign tournaments to VenueCategories in the
formset.

It also works on existing Venues and VenueCategories in a migration,
which assigns tournaments to venues with debates associated as well as
VenueCategories with venues associated, splitting them if they are
linked to many tournaments implictly.

For good measure, a second round of finding venues' tournaments is
conducted through VenueCategories.