Migrate participant ballots for concurrent rounds
With #516, there can be many rounds happening at the same time, but it
is not reflected in the public ballot submission views, of which
silently fail, showing a wrong ballot, if adjudicators are assigned to
concurrent rounds.

To fix, the views are now Round-based with the sequence in the URL, with
links split per-round.