Add ability to merge ballots
This commit adds a new admin/assistant page that takes each
adjudicator's most recent partial ballot for a debate then merges them,
leading to the ballot page to verify and submit the completed ballot.

This merge works even if ballots are missing, with the expectation that
missing ballots would be added by tab. However, if there are
discrepancies in adjudicators selecting different motion/vetos or have
speakers in a different order, the merge is canceled, leading back to
the results list with a message explaining the error. This failure
protects the scores which may themselves be in the wrong order.