Pre-populate speakers in public ballot submissions
To minimise needing to re-enter speaker order by each adjudicator in a
debate, if there is an order that already exists, use that.

This commit populates a DebateResult with speakers to initialize a
ballot form. As this is pre-filling data which may be erroneous, the
scope is limited to only partial ballots by a different submitter,
preferring confirmed/newer ballots. A warning is also added to double-
check the values.

Not just reducing time for submitters, but this may reduce failed merges
as there is a compatible suggestion provided.

As an extra, motions/vetos are also populated.