Restore silent failure of a29981d5
This restores the line added in a29981d58bc334100cf1cd3946d27a1e5f69ce7c
to make set_score() fail silently if the adjudicator is not on the
panel. It was removed by the merger of #1180, specifically in
a0eb1f7246df74b5a941219bd2737185550f884d, which had been done long
before a29981d5 but not been merged until it was merged with #1180
(i.e., it was a merge conflict whose resolution appears to have
been overlooked).

Fixes #922.