Add preference to exclude ghosts from team scores
As pointed out recently, the WUDC and EUDC rules show that team scores
must not include scores from iron-men (except for their highest). This
is 34.9.3.2 in WUDC. The current behaviour was to keep ghost scores, and
is expected now. As such, a new preference is added to toggle between
these calculations, but is set correctly in the preset.