Show None values in Standings tables
This commit creates mechanisms to handle None values in standings,
which can occur if there is not enough data for an aggregation for a
participant. Instead of showing 0s in those cases, a hyphen is shown
in place, like with missing rounds.

To handle Nones, a "default" value of 0 for sorting was set in the
metric getter.

This commit also adds SQL-based tiebreak functions when specific to
types of participant.