Use SQL migration for ManyAnswer
As we've been getting errors that we cannot ALTER TABLE
"adjfeedback_adjudicatorfeedbackquestion" because it has pending trigger
events, the migration that caused the error should be checked.

The error is probably caused by running Python on an un-finished
migration, and thus cannot done in a single SQL transaction. This
commit converts the Python routine to an SQL command to re-establish the
transaction.