Fix N+1 queryset issue in feedback progress
A line in expected_feedback_targets() bypassed the prefetched queryset, because
it used .first() on a nondeterministic queryset to check a property of the
first element of the queryset. This caused an N+1 issue on the debate
adjudicators for each debate.

Related (rejected) Django ticket explaining why this ignores the prefetch:
https://code.djangoproject.com/ticket/23805

The line in question was added in 51092e9261d8a6a8515bc0388196fa714b9f45c3
as part of #1387.