Use form_valid for notification forms, not post
As safety so that the form won't submit and try to send message(s) if
the form was invalid. Now the raw POST params are less called (although
they still need to be for the recipients' list as that is not handled
by the form itself)

Also added a check that the provided template is suitable and can be
interpreted by Django before submitting (thus as a ValidationError).
(Think using lambdas for the cleaning functions would have been cleaner
as code)

Fixes #1812