Remove Donation pages in favour of OpenCollective
Instead of having a donation page that makes a plea and redirects to
another platform, have that platform have the plea and the payment on
the same page to reduce friction. The footer also has a direct link to a
payment page with the number of billable teams where applicable.