API: Create availabilities endpoints
This commit creates and documents a new API endpoint to set and unset
availabilities for specific rounds. Here, we treat availabilities as a
list of objects that are available in the round as there isn't other
data to worry about. This does mean that we treat the single list field
as the serializer itself.

The list is by default empty, expecting query parameters to specify what
type of participant to include. This makes the DELETE verb safer and a
smaller response. The same verbs are used for availabilities as for
check-ins, but the POST verb is the equivalent of the checkin DELETE,
as there is no "creation" and the DELETE verb does not take a request
body.