Make Calico deployment more prominent
This commit moves the paragraph about Calico over Heroku and adds a
similar "Deploy" button. It also fixes a broken anchor.

Calico should be prioritized for a few reasons, namely that it is the
most effective tool to enforce the licence, and that it requires the
least effort for the users to deploy and maintain, with fewer problems
arising.

The link to create a site in the footer is also replaced with the README
as to present Calico.