Add DEFAULT_AUTO_FIELD setting; closes #1810
Django 3.2 introduces this setting and makes the default AutoField, but gives
lots of annoying warnings in the logs. This sets the setting explicitly to
mute the warnings, and also to avoid unwanted migrations when in a future
version of Django the default is changed to BigAutoField.

See commentary in #1810, and also:
https://docs.djangoproject.com/en/3.2/releases/3.2/#customizing-type-of-auto-created-primary-keys