Add field for Tabbycat version in API
The field is on the root endpoint, and shows the current Tabbycat
version, which would indicate to applications which features in the API
can be used (we only ensure backwards compatibility, not forwards).

At the same time, update the docs with the root schema, and append
"-dev" to our version.