API: Add validation for unique together
As DRF cannot detect exceptions caused by unique together constraints
where some fields are not in the serializer (tournament), this commit
adds a free method that checks for the existence of such object and
raises a validation error on that field, with partialmethod. More
complex contraints are in the general validation method.