Remove GitHub reporting link
GitHub should be used for the development of Tabbycat, including in
making error reports. However, we can't really assume the technical
savvy of our users to create useful and relevant reports, so they end up
being closed out-of-hand. Posting them in the FB group would allow for
more eyes and better triage for what could be a true bug.

To try to make FB reports better, an instruction to state intentions is
also added in addition to the URL.