Create new field to obscure participant names
This commit creates a new field for speakers and adjudicators so that
they may have code names shown in public views. This feature would allow
for privacy in school tournaments and to avoid bias based on names.

Old participants will not have a code name assigned, but new People will
have a numerical identifier used as code name.