API: Create endpoints for Preformed Panels
This commit creates a viewset for the creation and handling of preformed
panels, attempting to use the polymorphism of AdjudicatorAllocation be-
tween Debates and Preformed Panels.

In addition to the regular CRUD methods, DELETE and PUT have been added
to the list endpoint, allowing for all preformed panels for the round to
be deleted, and for anticipated (blank) panels to be added, respectively.