Unset round sequence and draw status as editable
The Database area Rounds list has the sequence and draw status editable
right in the list. Editing sequences is liable to fail with a 500 error
as the sequence has a uniqueness contraint.

The draw status is dangerous to have modifyable in a list as the status
can change through the use of Tabbycat in ways that aren't immediately
reflected. If the page is saved when out-of-date, the statuses get
reverted, leading to confusion which can lead to data loss.

Fixes BACKEND-9MV