API: Avoid popping fields in documentation
This commit fixes errors found when trying to create a schema where
fields were being popped. Spectacular sometimes doesn't include the
"context" kwarg to the serializers, and is unauthenticated, which makes
the serializers pop sensitive fields.

As the host is always "testserver", we can test for that and give access
to the schema generator.