Add field for iron-person on TeamScore
This commit adds a new boolean field to TeamScore to indicate whether a
team had an iron-person speech. The migration adding the field also runs
that check to update existing TeamScores.

The field will be useful to make it easier to count the number of times
a team was an iron, which is often a break criterion.