Flesh out email context generators
As we are making them into classes, we might as well add their data-
classes as an attribute, and they can be class methods.

Then, the USER parameter could be made within the main email loop as the
Person object is the 2nd element of each tuple.

Also made a few other changes suggested by my IDE.