Regroup motion release messages to use plural
This commit consolidates messages in motion release to avoid "(s)" and
use the pluralization capabilities of gettext. So that messages indi-
cating the number of motions when none are entered show the correct
plural form, if motions are enabled it'll assume there are 3 motions.
Otherwise, it'll take the number of motions entered or 1 if none are.