API: Hide name when anonymous or using codes
This commit creates a new field type for person serializers (speaker &
adjudicator) which makes the name null if anonymous or their coded name
if enabled by the tournament. It does not apply to users with permission
though, and names can still be edited.

The "anonymous" field does not have to be elided as it'd be clear from
null values if a person is anonymous.