Add break size validation to Break Category form
This commit adds validation to the Break Category formset to prevent BP
tournaments from having a break category with an invalid size (full or
partial). Adding this validation required passing the tournament to the
newly-subclassed model form.

Further, the saving sequence is fixed so that the break rounds are
generated after saving the object.

The tournament field is now ignored as that information is found in the
view object. It is however kept for IntegrityError catching.

Closes BACKEND-AXB
Closes BACKEND-AMJ