Add team seeding for first round
This commit adds a "seed" attribute to Teams which are used for rankings
in the first round if power-pairing is used. In that situation, all
teams are placed in the same bracket with the Standings being manually
created. It can then use the normal power-pairing draw generator.

To continue this feature, a new conflict avoidance should be made to
minimize conflicts between pairings where teams have the same seed, i.e.

A 1 (seed 1) vs A 2 (seed 2)
B 1 (seed 1) vs B 2 (seed 2)

would become:
A 1 vs B 2
B 1 vs A 2

Fixes #2165