API: Set default round stage to preliminary
This commit sets the default round stage to Preliminary in the Round
serializer's break validation, as to allow preliminary rounds to be
created if not specified, in accordance with the model's default.

It also sets the `motion` field as optional, as it is popped with an
empty list default when treating it.