Create metric for number of times in pullup room
This metric was requested by APDA tournaments, which attempt to minimize
not only the number of times a team gets pulled up, but also the times
teams debate against pullups, as facing a pullup may confer an outsized
advantage to the stronger team. This metric is not expected to be used
in standings, only in draw generation.

It is named "number of times in pullup debates" and has a weird
abbreviation.

The commit also renames the "number of pullups" metric to remove the
"before this round" comment as extraneous.