Add penalty preference for seeing pullup again
This commit adds a new preference for the penalty to apply for re-seeing
a pullup. The preference is not yet "connected" to anything, but follows
the addition of the related "saw_pullups" metric.

Rather than creating two preferences, one boolean to activate and then
the integer, a penalty of 0 can signal that the penalty should not be
applied. Thus, only the penalty integer preference has been added (and
thus has a default of 0). The application of a preset that includes the
preference would be an indication that it can be customized.