Implement new draw algorithm for 2-team formats
This commit creates and implements a new algorithm for pairing teams in
two-team formats, which combines pairing methods and conflict avoidance.
This new method calculates a badness score for all possible pairings
within brackets and then minimizes the cost of making the pairings. This
is in contrast with the previous method which creates the pairings and
then makes swaps iteratively to avoid conflicts.

The method uses graph theory to create a graph where each node is a team
and the edges are possible pairings with a weight based on the penalty
given to the matchup. The graph is undirected (although if directed
could be used for side allocation), and then the Blossom algorithm for
maximal weight graph matching is used as its dual to minimize costs.
When sides are preallocated we use the same cost function but with the
Hungarian algorithm. This is possible as preallocating sides makes the
graph bipartite, a condition to use that algorithm.

The Blossom algorithm and Graph object class is provided by a new
package: networkx. Normally it requires SciPy with its default
installation, but the methods we use don't require them and so are
excluded.

The cost function has many components:
* History/institution conflicts, handled with the same penalty
  preferences as with one-up-one-down conflict avoidance.
* Side balance, handled with a new "side_penalty" preference, used when
  balance is requested. As matching is done in an undirected graph, we
  cannot make the costs different for which side a team is assigned.
  Thus, we calculate if the pairing would aggravate an existing
  imbalance and give a penalty on what the difference would become times
  the penalty to attribute.
* Pairing deviations, a small penalty given for not being paired with
  the team in the intended position, the difference of the intended
  matchup and the pair being the coefficient for the new
  "pairing_penalty" preference. (i.e. with 8 teams and folding, the 1-6
  pair would have a difference of 2)

Pullups are decided before the application of this method and so are not
affected.