Add "saw pullup" penalty for draw optimisation
This commit implements the "saw pullups" penalty preference for use in
the new draw optimisation algorithm. For it to work, the "saw pullups"
metric must be added to the standings generator, with the pullups
metric. With a non-zero penalty, the metric will be automatically
added.

Note that the "saw pullup" will also apply to the team who got pulled
up.