Create model for user permissions
This commit creates a new model for assigning specific permissions per
tournament to a user, with an addition to Edit Database for assigning
them.

To start, some permissions relating to editing or accessing conflicts
was added. The maximum length was put at 50 to allow for some expansion.
The current permissions follow a pattern of "edit"/"view".model for
consistency with the action and where.

In addition, a new helper function is created to determine whether a
given user has the requested permission. These permissions will only
apply to logged-in users as unauthenticated access to various parts of
the platform is already handled through settings. Further, superusers
will bypass permissions. This is to maintain backwards compatibility,
but also as a step towards deprecating the Assistant view (used by
authenticated users without advanced privileges) towards them having a
restricted access to the Admin interface.

As permissions may be frequently accessed, they'll be cached similarly
to tournaments in TournamentsMixin, using an attribute on the user, then
through caching, until fetching from the DB. There's also a wrapper for
the method, to be used within templates.